requirements.txt
Django==4.2.14
psycopg==3.2.1
psycopg-binary==3.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
pillow==10.4.0
drf-nested-routers==0.93.4
pytest-django==4.8.0
